Reimplement Electric.Postgres.LockBreakerConnection as a one-off query
